requirements.txt
google-generativeai
discord.py
jishaku
python-dotenv